requirements.txt
# Core Dependencies
python-dotenv==1.0.0
fastapi==0.109.0
uvicorn==0.27.0
streamlit==1.30.0

# LangChain and LLM
langchain-core>=0.3.49
langchain>=0.1.5
langchain-community>=0.0.20
langchain-groq>=0.3.2
groq>=0.4.1
pydantic>=2.0.0

# File Processing
python-multipart==0.0.6

# Vector Store and Search
opensearch-py==2.4.2

# Data Processing
pandas==2.1.4
numpy==1.26.3
Pillow==10.2.0

# Development Tools
black==23.12.1
pytest==7.4.4
